FIX: Sanitize parameters provided to user actions
Currently, providing things like `filter[%24acunetix]=1` to
`UserActionsController#index` will throw an exception because instead of
getting a string as expected, we get a hash instead.

This patch simply converts all the parameter values to strings so even
if the provided data makes no sense, we can manipulate it without
raising any error and the expected workflow will be applied.